renaming
Changed description and name

This was a fixed version of a StK map made by Primo. Although, it seems that nowadays Primo is actively working on their original version, so go play the original I guess, although currently I still prefer the old version, so maybe this would be a way for someone to enjoy more of that older version vibes (with some of my tweaks and changes).